prepared interpreter fix
