+++ .vscode/launch.json
@@ -8,7 +8,7 @@
             "name": "Python: Aktuelle Datei",
             "type": "python",
             "request": "launch",
-            "program": "run.py",
+            "program": "debug.py",
             "console": "integratedTerminal"
         }
     ]
